+++ FlowCrypt/Functionality/Mail Provider/Message Provider/Gmail+Message.swift
@@ -31,7 +31,7 @@ extension GmailService: MessageProvider {
                 }
 
                 do {
-                    let message = try Message(gmailMessage)
+                    let message = try Message(gmailMessage: gmailMessage)
                     return continuation.resume(returning: message)
                 } catch {
                     return continuation.resume(throwing: error)
@@ -130,144 +130,5 @@ extension GmailService: MessageProvider {
         let query = GTLRGmailQuery_UsersMessagesGet.query(withUserId: .me, identifier: identifier)
         query.format = format
         return query
-    }
-}
-
-extension GTLRGmail_Message {
-    var textParts: [GTLRGmail_MessagePart] {
-        payload?.parts?.filter { $0.filename.isEmptyOrNil } ?? []
-    }
-
-    var attachmentParts: [GTLRGmail_MessagePart] {
-        payload?.parts?.filter { !$0.filename.isEmptyOrNil } ?? []
-    }
-
-    func body(type: MessageBodyType) -> Data? {
-        if let text = textParts.findMessageBody(type: type)?.body?.data {
-            return GTLRDecodeWebSafeBase64(text)
-        } else if let multipartBody = textParts.findMultipartBody(),
-                  let text = multipartBody.parts?.findMessageBody(type: type)?.body?.data {
-            return GTLRDecodeWebSafeBase64(text)
-        } else if let body = payload?.body?.data {
-            return GTLRDecodeWebSafeBase64(body)
-        } else {
-            return nil
-        }
-    }
-}
-
-enum MessageBodyType: String {
-    case text = "text/plain",
-         html = "text/html",
-         multipartAlternative = "multipart/alternative",
-         multipartMixed = "multipart/mixed"
-}
-
-extension Array where Iterator.Element == GTLRGmail_MessagePart {
-    func findMultipartBody() -> GTLRGmail_MessagePart? {
-        let types = [MessageBodyType.multipartMixed, MessageBodyType.multipartAlternative].map(\.rawValue)
-        return first(where: {
-            guard let mimeType = $0.mimeType else { return false }
-            return types.contains(mimeType)
-        })
-    }
-
-    func findMessageBody(type: MessageBodyType) -> GTLRGmail_MessagePart? {
-        first(where: { $0.mimeType == type.rawValue })
-    }
-}
-
-extension Message {
-    init(
-        _ message: GTLRGmail_Message,
-        draftIdentifier: String? = nil
-    ) throws {
-        guard let payload = message.payload else {
-            throw GmailServiceError.missingMessagePayload
-        }
-
-        guard let messageHeaders = payload.headers else {
-            throw GmailServiceError.missingMessageInfo("headers")
-        }
-
-        guard let internalDate = message.internalDate as? Double else {
-            throw GmailServiceError.missingMessageInfo("date")
-        }
-
-        guard let identifier = message.identifier else {
-            throw GmailServiceError.missingMessageInfo("id")
-        }
-
-        let attachmentsIds = payload.parts?.compactMap { $0.body?.attachmentId } ?? []
-        let labels: [MessageLabel] = message.labelIds?.map(MessageLabel.init) ?? []
-        let bodyHtml = message.body(type: .html)?.toStr()
-        let bodyText = message.body(type: .text)?.toStr() ?? bodyHtml?.htmlToString() ?? ""
-        let body = MessageBody(text: bodyText, html: bodyHtml)
-        let attachments: [MessageAttachment] = message.attachmentParts.compactMap {
-            guard let body = $0.body, let id = body.attachmentId, let name = $0.filename
-            else { return nil }
-            return MessageAttachment(
-                id: Identifier(stringId: id),
-                name: name,
-                estimatedSize: body.size?.intValue,
-                mimeType: $0.mimeType,
-                data: nil
-            )
-        }
-
-        var sender: Recipient?
-        var subject: String?
-        var to: String?
-        var cc: String?
-        var bcc: String?
-        var replyTo: String?
-        var inReplyTo: String?
-
-        for messageHeader in messageHeaders.compactMap({ $0 }) {
-            guard let name = messageHeader.name?.lowercased(),
-                  let value = messageHeader.value
-            else { continue }
-
-            switch name {
-            case .from: sender = Recipient(value)
-            case .subject: subject = value
-            case .to: to = value
-            case .cc: cc = value
-            case .bcc: bcc = value
-            case .replyTo: replyTo = value
-            case .inReplyTo: inReplyTo = value
-            default: break
-            }
-        }
-
-        self.init(
-            identifier: Identifier(stringId: identifier),
-            // Should be divided by 1000, because Date(timeIntervalSince1970:) expects seconds
-            // but GTLRGmail_Message.internalDate is in miliseconds
-            date: Date(timeIntervalSince1970: internalDate / 1000),
-            sender: sender,
-            subject: subject,
-            size: message.sizeEstimate.flatMap(Int.init),
-            labels: labels,
-            attachmentIds: attachmentsIds,
-            body: body,
-            attachments: attachments,
-            threadId: message.threadId,
-            draftIdentifier: draftIdentifier,
-            raw: message.raw,
-            to: to,
-            cc: cc,
-            bcc: bcc,
-            replyTo: replyTo,
-            inReplyTo: inReplyTo
-        )
-    }
-}
-
-extension String {
-    func htmlToString() -> String? {
-        return try? NSAttributedString(data: self.data(using: .utf8)!,
-                                        options: [.documentType: NSAttributedString.DocumentType.html],
-                                        documentAttributes: nil).string
     }
 }
